+++ backend/connector-integration/src/connectors/cybersource/transformers.rs
@@ -3711,6 +3711,7 @@ impl From<&ClientProcessorInformation> for AdditionalPaymentMethodConnectorRespo
             payment_checks,
             card_network: None,
             domestic_network: None,
+            auth_code: None,
         }
     }
 }